connecting the server to the sqlite db
